fix: ensure we can build containers properly
It's very difficult to execute bash in a `Dockerfile` `RUN` line; as such, it's better too provide a script file that uses the `#!/bin/bash` line.
That said, you need to call it using the "exec" style of `RUN` (`RUN [...]`), as otherwise it calls `/bin/sh -c {script}`.

Using a script also simplifies the `Dockerfile`, as I can add all scripts at once, and then just call the ones needed to build the container.

Finally, I discovered some issues with the way the RUN script was done in terms of expanding package names, and fixed those issues.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>